Semi-Working Collission Resolving
Colliding cubes can now stand ontop of the plane. However because vertical motion is not stopped (due to gravity being constant), object proceeds to fall the next frame before being repelled in the one after, leading to the object vibrating in place.

Planned possible 'axis locking' to mimic the cube landing on a solid, impassable object (so it's axis will be locked from changing). Needs further research and testing