Decelerating Objects thanks to Friction
Objects can now decelerate due to the influence of Friction.

A review of the friction calculation revealed a logic error. Function was determining the rate of Friction based on gravity, and thus was determining air resistance which should be calculated by drag.

The function also returns the absolute value of friction. A further logic error was revealed that had friction being ADDED to the net force due to this misunderstanding (therefore propelling the object).

NEXT: Review Drag as same logic errors may be active.