Conditional Friction
A further logic error was revealed causing friction to apply to ANY force. This meant objects were decelerating in their descent to the floor - which should not happen at this stage due to no drag being applied.

A temp flag was introduced stating if the object was currently colliding. This allows the update function to determine if Friction or Drag should be applied to the current movement.

NEXT: Fix drag calculations
TODO:
- Fix Collision checks
- Allow Movement's _'isColliding' flag to be set by the collider when it determines a collision.