Working Drag
Feedback from Lecturers revealed a major logic error where the new velocity was being incorrectly calculated. (as acceleration was being multiplied by mass instead of by time).

This lead to the drag increasing exponentially due to the contrast in size between the object's mass and the faction of a second recorded by delta time (which works to keep the change in velocity small).

NEXT: Enable Static Friction Calculations