Lifted Objects now fall down
Objects propelled into the air by a force now fall back to the ground.

Bug was being caused by the friction negation causing gravity to all be removed, placing the object completely into stasus. Check was moved to run before gravity was added to ensure gravity is always being applied to an object