Changed Loaded Shapes
Changed loaded shapes from 1 sphere and 4 cubes to 2 sphere and 2 cubes. This makes it easier to demo sphere to sphere, AABB to AABB, and sphere to AABB collisions.

Also discovered a bug in the code where transforms for sphere colliders were not being set properly, leading collision checks to throw a 'NULL ERROR'.

Sphere - to - sphere collision was also checked as is confirmed to be working correctly.

NEXT: Create Sphere to AABB collision