Semi Working Collision Manifold
Refactored Sphere-to-Sphere Collision to populate a collision manifold which is now used to pass the collision's collision normal to the Movement component.

Code works as intended and functionality now needs to be expanded to AABB-to-AABB and Sphere-to-AABB.

Component relationships need to be looked at, as currently the Movement Component needs to access the entire collision component in order to access the collision manifold.